+++ CHANGELOG.md
@@ -9,6 +9,8 @@ New features:
 Behaviour changes:
  * adjust behaviour of ProcessPreBlock callback (#129)
  * (*DBFT).Header() and (*DBFT).PreHeader() are moved to (*Context) receiver (#133)
+ * support error handling for ProcessBlock callback if anti-MEV extension is enabled
+   (#134)
 
 Improvements:
  * minimum required Go version is 1.22 (#122, #126)